+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 - Bug #388: Set empty `comment` and `extra` properties to `null` when loading table columns (@Tigrov)
 - Enh #389, #390: Use `DbArrayHelper::arrange()` instead of `DbArrayHelper::index()` method (@Tigrov)
 - New #387: Realize `Schema::loadResultColumn()` method (@Tigrov)
+- New #393: Use `DateTimeColumn` class for datetime column types (@Tigrov)
 
 ## 1.2.0 March 21, 2024